[Refacto] Move getEnv, getPrismicData, validateUserAuth to server services directory
