+++ packages/slice-machine/server/src/api/auth/validate.ts
@@ -1,10 +1,11 @@
-import { validate } from "@lib/env/client";
+import { validateUserAuth } from "../services/validateUserAuth";
 
 type ValidateResponse = {
   connected: boolean;
   reason: string;
   err?: any;
 };
+
 export default async function handler(): Promise<ValidateResponse> {
-  return validate();
+  return validateUserAuth();
 }
