+++ packages/slice-machine/server/src/api/env.ts
@@ -1,4 +1,4 @@
-import { getEnv } from "@lib/env";
+import getEnv from "./services/getEnv";
 
 export default async function handler() {
   return getEnv();
